requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
asgiref==3.5.2
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6